tests/requirements.txt
pytest
requests
psycopg2-binary
factory_boy
python-dotenv
